requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
pip==23.1.2
pytz==2023.3
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.3